+++ src/main/java/nextstep/qna/exception/CannotDeleteException.java
@@ -1,4 +1,4 @@
-package nextstep.qna;
+package nextstep.qna.exception;
 
 public class CannotDeleteException extends Exception {
     private static final long serialVersionUID = 1L;
