+++ src/main/java/nextstep/qna/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.qna;
+package nextstep.qna.exception;
 
 public class NotFoundException extends RuntimeException {
 }
